requirements.txt
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv
streamlit
uvicorn
pydantic
httpx
requests
langchain_tavily<0.2
langchain_groq
langchain_openai
langgraph
ipykernel

-e .